Fixed sources localisation in finder
